[refactor] 제미나이 코드리뷰 기반 코드 수정 및 입금관련 기능 마무리

+++ src/main/java/dev/woori/wooriBank/domain/transaction/transfer/dto/TransferResponseDto.java
@@ -1,7 +1,6 @@
 package dev.woori.wooriBank.domain.transaction.transfer.dto;
 
 import lombok.Builder;
-import lombok.Getter;
 
 /**
  * 입금 처리 후 클라이언트에게 응답하는 DTO
@@ -12,7 +11,7 @@ import lombok.Getter;
 public record TransferResponseDto (
         String fromAccount, // 보내는 계좌 (관리자 계좌)
         String toAccount, // 받는 계좌 (사용자 계좌)
-        Integer amount, // 이체 금액
+        Long amount, // 이체 금액
         String message // 성공 메세지 또는 안내문구
 ){
